pyproject.toml
[project]
name = ""
version = "0.1.0"
description = ""
requires-python = ">=3.11"
dependencies = [
    "gymnasium",
    "pillow",
    "torch",
    "pyyaml",
    "wandb",
]